유저 리프레시 토큰 저장 upsert 문으로 변경(jpa query 적용)

+++ src/main/java/com/example/treeze/entity/login/User.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userSeq;
+    private Long userSeq;
 
     @Column(name = "user_id", length = 20, nullable = false)
     private String userId;
